system-backend/services/rag-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
pydantic==2.8.0
pydantic-settings==2.4.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy[asyncio]==2.0.35
confluent-kafka==2.5.0
sentence-transformers==3.1.1
torch==2.4.1
transformers==4.44.2
huggingface-hub==0.24.7
llama-cpp-python==0.2.90
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.2
structlog==24.4.0
